+++ modules/effectie-cats-effect2/shared/src/main/scala-2/effectie/ce2/canCatch.scala
@@ -1,8 +1,6 @@
 package effectie.ce2
 
-import cats.Id
 import cats.effect.IO
-import cats.syntax.all._
 import effectie.core.CanCatch
 
 /** @author Kevin Lee
@@ -16,24 +14,6 @@ object canCatch {
 
     @inline override def catchNonFatalThrowable[A](fa: => IO[A]): IO[Either[Throwable, A]] =
       fa.attempt
-
-  }
-
-  implicit object canCatchId extends CanCatch[Id] {
-
-    @inline override final def mapFa[A, B](fa: Id[A])(f: A => B): Id[B] = f(fa)
-
-    @inline override def catchNonFatalThrowable[A](fa: => Id[A]): Id[Either[Throwable, A]] =
-      scala.util.Try(fa) match {
-        case scala.util.Success(a) =>
-          a.asRight[Throwable]
-
-        case scala.util.Failure(scala.util.control.NonFatal(ex)) =>
-          ex.asLeft[A]
-
-        case scala.util.Failure(ex) =>
-          throw ex // scalafix:ok DisableSyntax.throw
-      }
 
   }
