+++ modules/effectie-cats-effect2/shared/src/main/scala-2/effectie/ce2/canHandleError.scala
@@ -1,10 +1,7 @@
 package effectie.ce2
 
-import cats.Id
 import cats.effect.IO
 import effectie.core.CanHandleError
-
-import scala.util.control.NonFatal
 
 /** @author Kevin Lee
   * @since 2020-08-17
@@ -18,20 +15,6 @@ object canHandleError {
 
     @inline override final def handleNonFatal[A, AA >: A](fa: => IO[A])(handleError: Throwable => AA): IO[AA] =
       handleNonFatalWith[A, AA](fa)(err => IO.pure(handleError(err)))
-
-  }
-
-  implicit object idCanHandleError extends CanHandleError[Id] {
-
-    @inline override final def handleNonFatalWith[A, AA >: A](fa: => Id[A])(handleError: Throwable => Id[AA]): Id[AA] =
-      try (fa)
-      catch {
-        case NonFatal(ex) =>
-          handleError(ex)
-      }
-
-    @inline override final def handleNonFatal[A, AA >: A](fa: => Id[A])(handleError: Throwable => AA): Id[AA] =
-      handleNonFatalWith[A, AA](fa)(err => handleError(err))
 
   }
