+++ modules/effectie-cats-effect2/shared/src/main/scala-2/effectie/ce2/fromFuture.scala
@@ -1,11 +1,9 @@
 package effectie.ce2
 
-import cats.Id
 import cats.effect.{ContextShift, IO}
 import effectie.core.FromFuture
-import effectie.instances.future.fromFuture.FromFutureToIdTimeout
 
-import scala.concurrent.{Await, Future}
+import scala.concurrent.Future
 
 /** @author Kevin Lee
   * @since 2020-09-22
@@ -17,13 +15,6 @@ object fromFuture {
     new FromFuture[IO] {
       override def toEffect[A](future: => Future[A]): IO[A] =
         IO.fromFuture[A](IO(future))
-    }
-
-  @SuppressWarnings(Array("org.wartremover.warts.ImplicitParameter"))
-  implicit def fromFutureToId(implicit timeout: FromFutureToIdTimeout): FromFuture[Id] =
-    new FromFuture[Id] {
-      override def toEffect[A](future: => Future[A]): Id[A] =
-        Await.result[A](future, timeout.fromFutureToIdTimeout)
     }
 
 }
