+++ modules/effectie-cats-effect2/shared/src/main/scala-2/effectie/ce2/toFuture.scala
@@ -1,10 +1,9 @@
 package effectie.ce2
 
-import cats.Id
 import cats.effect._
 import effectie.core.ToFuture
 
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.Future
 
 /** @author Kevin Lee
   * @since 2020-09-23
@@ -16,12 +15,5 @@ object toFuture {
     override def unsafeToFuture[A](fa: IO[A]): Future[A] =
       fa.unsafeToFuture()
   }
-
-  @SuppressWarnings(Array("org.wartremover.warts.ImplicitParameter"))
-  implicit def idToFuture(implicit executionContext: ExecutionContext): ToFuture[Id] =
-    new ToFuture[Id] {
-      override def unsafeToFuture[A](fa: Id[A]): Future[A] =
-        Future(fa)
-    }
 
 }
