+++ modules/effectie-cats-effect2/shared/src/test/scala-2/effectie/ce2/fxCtorSpec.scala
@@ -77,19 +77,22 @@ object fxCtorSpec extends Properties {
 
   private val futureSpecs = effectie.core.FxCtorSpec.futureSpecs
 
-  private val idSpecs = List(
-    property("test FxCtor[Id].effectOf", IdSpecs.testEffectOf),
-    property("test FxCtor[Id].pureOf", IdSpecs.testPureOf),
-    property("test FxCtor[Id].pureOrError(success case)", IdSpecs.testPureOrErrorSuccessCase),
-    example("test FxCtor[Id].pureOrError(error case)", IdSpecs.testPureOrErrorErrorCase),
-    example("test FxCtor[Id].unitOf", IdSpecs.testUnitOf),
-    example("test FxCtor[Id].errorOf", IdSpecs.testErrorOf),
-    property("test FxCtor[Id].fromEither(Right)", IdSpecs.testFromEitherRightCase),
-    property("test FxCtor[Id].fromEither(Left)", IdSpecs.testFromEitherLeftCase),
-    property("test FxCtor[Id].fromOption(Some)", IdSpecs.testFromOptionSomeCase),
-    property("test FxCtor[Id].fromOption(None)", IdSpecs.testFromOptionNoneCase),
-    property("test FxCtor[Id].fromTry(Success)", IdSpecs.testFromTrySuccessCase),
-    property("test FxCtor[Id].fromTry(Failure)", IdSpecs.testFromTryFailureCase),
-  )
+  private val idSpecs = {
+    import effectie.instances.id.fxCtor._
+    List(
+      property("test FxCtor[Id].effectOf", IdSpecs.testEffectOf),
+      property("test FxCtor[Id].pureOf", IdSpecs.testPureOf),
+      property("test FxCtor[Id].pureOrError(success case)", IdSpecs.testPureOrErrorSuccessCase),
+      example("test FxCtor[Id].pureOrError(error case)", IdSpecs.testPureOrErrorErrorCase),
+      example("test FxCtor[Id].unitOf", IdSpecs.testUnitOf),
+      example("test FxCtor[Id].errorOf", IdSpecs.testErrorOf),
+      property("test FxCtor[Id].fromEither(Right)", IdSpecs.testFromEitherRightCase),
+      property("test FxCtor[Id].fromEither(Left)", IdSpecs.testFromEitherLeftCase),
+      property("test FxCtor[Id].fromOption(Some)", IdSpecs.testFromOptionSomeCase),
+      property("test FxCtor[Id].fromOption(None)", IdSpecs.testFromOptionNoneCase),
+      property("test FxCtor[Id].fromTry(Success)", IdSpecs.testFromTrySuccessCase),
+      property("test FxCtor[Id].fromTry(Failure)", IdSpecs.testFromTryFailureCase),
+    )
+  }
 
 }
