+++ modules/effectie-cats-effect2/shared/src/test/scala-2/effectie/syntax/fxSpec.scala
@@ -354,6 +354,7 @@ object fxSpec extends Properties {
   }
 
   object IdSpec {
+    import effectie.instances.id.fx._
 
     def testAll: Property = for {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
