+++ modules/effectie-cats-effect2/shared/src/test/scala-3/effectie/ce2/fxSpec.scala
@@ -15,6 +15,7 @@ import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog.*
 import hedgehog.runner.*
 import effectie.specs.fxSpec.{FxSpecs, IdSpecs}
+import effectie.specs.MonadSpec
 
 import scala.concurrent.Await
 import scala.util.control.{ControlThrowable, NonFatal}
