+++ modules/effectie-cats/shared/src/test/scala/effectie/instances/id/canCatchSpec.scala
@@ -1,15 +1,13 @@
-package effectie.ce2
+package effectie.instances.id
 
 import canCatch._
 import cats._
 import cats.data.EitherT
-import cats.effect._
 import cats.instances.all._
 import cats.syntax.all._
 import effectie.SomeControlThrowable
 import effectie.core._
-import effectie.syntax.error._
-import effectie.syntax.fx._
+import effectie.syntax.all._
 import effectie.testing.types._
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
@@ -26,67 +24,7 @@ object canCatchSpec extends Properties {
 
   private implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
 
-  override def tests: List[Test] = ioSpecs ++ futureSpecs ++ idSpecs
-
-  val ioSpecs = List(
-    /* IO */
-    example(
-      "test CanCatch[IO]catchNonFatalThrowable should catch NonFatal",
-      IoSpec.testCanCatch_IO_catchNonFatalThrowableShouldCatchNonFatal
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatalThrowable should not catch Fatal",
-      IoSpec.testCanCatch_IO_catchNonFatalThrowableShouldNotCatchFatal
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatalThrowable should return the successful result",
-      IoSpec.testCanCatch_IO_catchNonFatalThrowableShouldReturnSuccessfulResult
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatal should catch NonFatal",
-      IoSpec.testCanCatch_IO_catchNonFatalShouldCatchNonFatal
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatal should not catch Fatal",
-      IoSpec.testCanCatch_IO_catchNonFatalShouldNotCatchFatal
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatal should return the successful result",
-      IoSpec.testCanCatch_IO_catchNonFatalShouldReturnSuccessfulResult
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatalEither should catch NonFatal",
-      IoSpec.testCanCatch_IO_catchNonFatalEitherShouldCatchNonFatal
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatalEither should not catch Fatal",
-      IoSpec.testCanCatch_IO_catchNonFatalEitherShouldNotCatchFatal
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatalEither should return the successful result",
-      IoSpec.testCanCatch_IO_catchNonFatalEitherShouldReturnSuccessfulResult
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatalEither should return the failed result",
-      IoSpec.testCanCatch_IO_catchNonFatalEitherShouldReturnFailedResult
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatalEitherT should catch NonFatal",
-      IoSpec.testCanCatch_IO_catchNonFatalEitherTShouldCatchNonFatal
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatalEitherT should not catch Fatal",
-      IoSpec.testCanCatch_IO_catchNonFatalEitherTShouldNotCatchFatal
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatalEitherT should return the successful result",
-      IoSpec.testCanCatch_IO_catchNonFatalEitherTShouldReturnSuccessfulResult
-    ),
-    example(
-      "test CanCatch[IO]catchNonFatalEitherT should return the failed result",
-      IoSpec.testCanCatch_IO_catchNonFatalEitherTShouldReturnFailedResult
-    ),
-  )
+  override def tests: List[Test] = futureSpecs ++ idSpecs
 
   /* Future */
   val futureSpecs = effectie.core.CanCatchSpec.futureSpecs ++ List(
@@ -169,183 +107,7 @@ object canCatchSpec extends Properties {
     throw throwable // scalafix:ok DisableSyntax.throw
 
   def run[F[*]: FxCtor: Functor, A](a: => A): F[A] =
-    effectOf[F](a)
-
-  object IoSpec {
-
-    def testCanCatch_IO_catchNonFatalThrowableShouldCatchNonFatal: Result = {
-
-      val expectedExpcetion = new RuntimeException("Something's wrong")
-      val fa                = run[IO, Int](throwThrowable[Int](expectedExpcetion))
-      val expected          = expectedExpcetion.asLeft[Int]
-      val actual            = CanCatch[IO].catchNonFatalThrowable(fa).unsafeRunSync()
-
-      actual ==== expected
-    }
-
-    @SuppressWarnings(Array("org.wartremover.warts.ToString"))
-    def testCanCatch_IO_catchNonFatalThrowableShouldNotCatchFatal: Result = {
-
-      val fatalExpcetion = SomeControlThrowable("Something's wrong")
-      val fa             = run[IO, Int](throwThrowable[Int](fatalExpcetion))
-
-      try {
-        val actual = CanCatch[IO].catchNonFatalThrowable(fa).unsafeRunSync()
-        Result.failure.log(s"The expected fatal exception was not thrown. actual: ${actual.toString}")
-      } catch {
-        case ex: SomeControlThrowable =>
-          ex.getMessage ==== fatalExpcetion.getMessage
-
-        case ex: Throwable =>
-          Result.failure.log(s"Unexpected Throwable: ${ex.toString}")
-      }
-
-    }
-
-    def testCanCatch_IO_catchNonFatalThrowableShouldReturnSuccessfulResult: Result = {
-
-      val fa       = run[IO, Int](1)
-      val expected = 1.asRight[Throwable]
-      val actual   = CanCatch[IO].catchNonFatalThrowable(fa).unsafeRunSync()
-
-      actual ==== expected
-    }
-
-    def testCanCatch_IO_catchNonFatalShouldCatchNonFatal: Result = {
-
-      val expectedExpcetion = new RuntimeException("Something's wrong")
-      val fa                = run[IO, Int](throwThrowable[Int](expectedExpcetion))
-      val expected          = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-      val actual            = CanCatch[IO].catchNonFatal(fa)(SomeError.someThrowable).unsafeRunSync()
-
-      actual ==== expected
-    }
-
-    @SuppressWarnings(Array("org.wartremover.warts.ToString"))
-    def testCanCatch_IO_catchNonFatalShouldNotCatchFatal: Result = {
-
-      val fatalExpcetion = SomeControlThrowable("Something's wrong")
-      val fa             = run[IO, Int](throwThrowable[Int](fatalExpcetion))
-
-      try {
-        val actual = CanCatch[IO].catchNonFatal(fa)(SomeError.someThrowable).unsafeRunSync()
-        Result.failure.log(s"The expected fatal exception was not thrown. actual: ${actual.toString}")
-      } catch {
-        case ex: SomeControlThrowable =>
-          ex.getMessage ==== fatalExpcetion.getMessage
-
-        case ex: Throwable =>
-          Result.failure.log(s"Unexpected Throwable: ${ex.toString}")
-      }
-
-    }
-
-    def testCanCatch_IO_catchNonFatalShouldReturnSuccessfulResult: Result = {
-
-      val fa       = run[IO, Int](1)
-      val expected = 1.asRight[SomeError]
-      val actual   = CanCatch[IO].catchNonFatal(fa)(SomeError.someThrowable).unsafeRunSync()
-
-      actual ==== expected
-    }
-
-    def testCanCatch_IO_catchNonFatalEitherShouldCatchNonFatal: Result = {
-
-      val expectedExpcetion = new RuntimeException("Something's wrong")
-      val fa                = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
-      val expected          = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-      val actual            = CanCatch[IO].catchNonFatalEither(fa)(SomeError.someThrowable).unsafeRunSync()
-
-      actual ==== expected
-    }
-
-    @SuppressWarnings(Array("org.wartremover.warts.ToString"))
-    def testCanCatch_IO_catchNonFatalEitherShouldNotCatchFatal: Result = {
-
-      val fatalExpcetion = SomeControlThrowable("Something's wrong")
-      val fa             = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](fatalExpcetion))
-
-      try {
-        val actual = CanCatch[IO].catchNonFatalEither(fa)(SomeError.someThrowable).unsafeRunSync()
-        Result.failure.log(s"The expected fatal exception was not thrown. actual: ${actual.toString}")
-      } catch {
-        case ex: SomeControlThrowable =>
-          ex.getMessage ==== fatalExpcetion.getMessage
-
-        case ex: Throwable =>
-          Result.failure.log(s"Unexpected Throwable: ${ex.toString}")
-      }
-
-    }
-
-    def testCanCatch_IO_catchNonFatalEitherShouldReturnSuccessfulResult: Result = {
-
-      val fa       = run[IO, Either[SomeError, Int]](1.asRight[SomeError])
-      val expected = 1.asRight[SomeError]
-      val actual   = CanCatch[IO].catchNonFatalEither(fa)(SomeError.someThrowable).unsafeRunSync()
-
-      actual ==== expected
-    }
-
-    def testCanCatch_IO_catchNonFatalEitherShouldReturnFailedResult: Result = {
-
-      val expectedFailure = SomeError.message("Failed")
-      val fa              = run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int])
-      val expected        = expectedFailure.asLeft[Int]
-      val actual          = CanCatch[IO].catchNonFatalEither(fa)(SomeError.someThrowable).unsafeRunSync()
-
-      actual ==== expected
-    }
-
-    def testCanCatch_IO_catchNonFatalEitherTShouldCatchNonFatal: Result = {
-
-      val expectedExpcetion = new RuntimeException("Something's wrong")
-      val fa       = EitherT(run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion)))
-      val expected = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-      val actual   = CanCatch[IO].catchNonFatalEitherT(fa)(SomeError.someThrowable).value.unsafeRunSync()
-
-      actual ==== expected
-    }
-
-    @SuppressWarnings(Array("org.wartremover.warts.ToString"))
-    def testCanCatch_IO_catchNonFatalEitherTShouldNotCatchFatal: Result = {
-
-      val fatalExpcetion = SomeControlThrowable("Something's wrong")
-      val fa = EitherT(run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](fatalExpcetion)))
-
-      try {
-        val actual = CanCatch[IO].catchNonFatalEitherT(fa)(SomeError.someThrowable).value.unsafeRunSync()
-        Result.failure.log(s"The expected fatal exception was not thrown. actual: ${actual.toString}")
-      } catch {
-        case ex: SomeControlThrowable =>
-          ex.getMessage ==== fatalExpcetion.getMessage
-
-        case ex: Throwable =>
-          Result.failure.log(s"Unexpected Throwable: ${ex.toString}")
-      }
-
-    }
-
-    def testCanCatch_IO_catchNonFatalEitherTShouldReturnSuccessfulResult: Result = {
-
-      val fa       = EitherT(run[IO, Either[SomeError, Int]](1.asRight[SomeError]))
-      val expected = 1.asRight[SomeError]
-      val actual   = CanCatch[IO].catchNonFatalEitherT(fa)(SomeError.someThrowable).value.unsafeRunSync()
-
-      actual ==== expected
-    }
-
-    def testCanCatch_IO_catchNonFatalEitherTShouldReturnFailedResult: Result = {
-
-      val expectedFailure = SomeError.message("Failed")
-      val fa              = EitherT(run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int]))
-      val expected        = expectedFailure.asLeft[Int]
-      val actual          = CanCatch[IO].catchNonFatalEitherT(fa)(SomeError.someThrowable).value.unsafeRunSync()
-
-      actual ==== expected
-    }
-
-  }
+    FxCtor[F].effectOf(a)
 
   object FutureSpec {
     import effectie.instances.future.canCatch._
@@ -525,8 +287,6 @@ object canCatchSpec extends Properties {
   }
 
   object IdSpec {
-    import effectie.instances.id.fxCtor._
-    import effectie.instances.id.canCatch._
 
     def testCanCatch_Id_catchNonFatalThrowableShouldCatchNonFatal: Result = {
