+++ modules/effectie-cats/shared/src/test/scala/effectie/instances/id/fromFutureSpec.scala
@@ -1,8 +1,6 @@
-package effectie.ce2
+package effectie.instances.id
 
 import cats.Id
-import cats.effect._
-import effectie.ce2.compat.CatsEffectIoCompatForFuture
 import effectie.core.FromFuture
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
@@ -11,6 +9,7 @@ import effectie.instances.future.fromFuture
 import hedgehog._
 import hedgehog.runner._
 
+import java.util.concurrent.ExecutorService
 import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -23,20 +22,24 @@ object fromFutureSpec extends Properties {
   private val waitFor300Millis = WaitFor(300.milliseconds)
 
   override def tests: List[Test] = List(
-    property("test FromFuture[IO].toEffect", IoSpec.testToEffect),
+    property("test FromFuture[Future].toEffect", FutureSpec.testToEffect),
     property("test FromFuture[Id].toEffect", IdSpec.testToEffect)
   )
 
-  object IoSpec {
+  object FutureSpec {
+
+    import effectie.instances.future.fromFuture._
+
     def testToEffect: Property = for {
       a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
     } yield {
-      val compat = new CatsEffectIoCompatForFuture
-      import compat._
+      implicit val es: ExecutorService  = ConcurrentSupport.newExecutorService(2)
+      implicit val ec: ExecutionContext =
+        ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
 
       ConcurrentSupport.runAndShutdown(es, waitFor300Millis) {
         lazy val fa = Future(a)
-        val actual  = FromFuture[IO].toEffect(fa).unsafeRunSync()
+        val actual  = ConcurrentSupport.futureToValueAndTerminate(es, waitFor300Millis)(FromFuture[Future].toEffect(fa))
 
         actual ==== a
       }
@@ -44,8 +47,6 @@ object fromFutureSpec extends Properties {
   }
 
   object IdSpec {
-    import effectie.instances.id.fromFuture._
-
     def testToEffect: Property = for {
       a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
     } yield {
