+++ modules/effectie-cats/shared/src/test/scala/effectie/instances/id/toFutureSpec.scala
@@ -1,17 +1,16 @@
-package effectie.ce2
+package effectie.instances.id
 
 import cats._
-import cats.effect._
-import toFuture._
 import effectie.core.ToFuture
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog._
 import hedgehog.runner._
+import toFuture._
 
 import java.util.concurrent.ExecutorService
-import scala.concurrent.Future
 import scala.concurrent.duration._
+import scala.concurrent.{ExecutionContext, Future}
 
 /** @author Kevin Lee
   * @since 2020-09-23
@@ -21,8 +20,8 @@ object toFutureSpec extends Properties {
 
   override def tests: List[Test] = List(
     property(
-      "test ToFuture[IO].unsafeToFuture",
-      IoSpec.testUnsafeToFuture
+      "test ToFuture[Future].unsafeToFuture",
+      FutureSpec.testUnsafeToFuture
     ),
     property(
       "test ToFuture[Id].unsafeToFuture",
@@ -30,7 +29,7 @@ object toFutureSpec extends Properties {
     )
   )
 
-  object IoSpec {
+  object FutureSpec {
     private val waitFor800Millis = WaitFor(800.milliseconds)
 
     @SuppressWarnings(Array("org.wartremover.warts.IsInstanceOf"))
@@ -38,34 +37,25 @@ object toFutureSpec extends Properties {
       a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
     } yield {
       val expected = a
-      val fa       = IO(expected)
 
-      implicit val es: ExecutorService = ConcurrentSupport.newExecutorService(2)
+      implicit val es: ExecutorService  = ConcurrentSupport.newExecutorService(2)
       @SuppressWarnings(Array("org.wartremover.warts.ExplicitImplicitTypes"))
-      implicit val ec                  =
+      implicit val ec: ExecutionContext =
         ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
       ConcurrentSupport.runAndShutdown(es, waitFor800Millis) {
-        val future   = ToFuture[IO].unsafeToFuture(fa)
-        val ioResult = fa.unsafeRunSync() ==== expected
-        val actual   = ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(500.milliseconds))(future)
+        val fa     = Future(expected)
+        import effectie.instances.future.toFuture._
+        val future = ToFuture[Future]
+        val actual =
+          ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(500.milliseconds))(future.unsafeToFuture(fa))
 
-        Result.all(
-          List(
-            ioResult,
-            Result
-              .assert(future.isInstanceOf[Future[Int]]) // scalafix:ok DisableSyntax.isInstanceOf
-              .log(s"future is not an instance of Future[Int]. future.getClass: ${future.getClass.toString}"),
-            actual ==== expected
-          )
-        )
+        actual ==== expected
       }
     }
 
   }
 
   object IdSpec {
-    import effectie.instances.id.toFuture._
-
     private val waitFor300Millis = WaitFor(300.milliseconds)
 
     @SuppressWarnings(Array("org.wartremover.warts.IsInstanceOf"))
