+++ modules/effectie-monix3/shared/src/test/scala/effectie/monix3/canCatchSpec.scala
@@ -762,7 +762,7 @@ object canCatchSpec extends Properties {
   }
 
   object IdSpec {
-    import effectie.ce2.fxCtor.idFxCtor
+    import effectie.instances.id.fxCtor._
 
     def testCanCatch_Id_catchNonFatalThrowableShouldCatchNonFatal: Result = {
