+++ modules/effectie-monix3/shared/src/test/scala/effectie/syntax/fxSpec.scala
@@ -401,7 +401,7 @@ object fxSpec extends Properties {
   }
 
   object IdSpec {
-    import effectie.ce2.fx.idFx
+    import effectie.instances.id.fx._
 
     def testAll: Property = for {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
